Revert "Simplify qml resource path"
This reverts commit d4e68fc1e164940dd30666d565837e5c98780f89.

This doesn't seem to work correctly right now. When copying the binary
to another place, it can't load different QML files, suggesting that
some relative paths are needed, which would be a downgrade from our
previous setup where we just needed to install the binary and where all
QML files were embedded as a resource in the binary.