Scale icon sizes
According to the Qt docs [0] for theme icons (which is the case here),
the closest available size will be chosen. This means that if we are
unlucky with just the right/wrong scaling factor, the chosen size will
be suboptimal. Instead we now force a larger size to avoid small icons
on HighDPI screens.

[0] https://doc.qt.io/qt-6/qtquickcontrols-icons.html#using-icons