Improve default height
If no items are visible, the height doesn't change from the current
height.

If items are visible, then the height automatically adapts to the
content height of the ListView, but it is kept between a reasonable
minimum and maximum height.

One remaining bug is that the ToolTip can steal mouse events. Hover
events have been fixed before [0], but it may still steal clicks, which
isn't as straight forward to fix, because qquicktooltip.cpp doesn't have
a function to disable or pass through click events.

[0] https://bugreports.qt.io/browse/QTBUG-63644