Disable tooltips for only one item
There is a bug where the tooltip steals mouse events and unfortunately
for only one item the size of the window is small enough that the
tooltip appears directly below the cursor (tooltips cannot be placed
outside of the window).

As a workaround disable tooltips completely in this edge case, they
are not that useful for only one item anyway.