Add support for drag notifications
Some notifications servers such as the default KDE Plasma notification
server allow sending a notification with a hint that specifies URIs to
be dragged, if the user attempts to drag the notification.

This patch adds support for utilizing this notification system as a
frontend instead of the default window frontend.

Something to be kept in mind is that KDE Plasma won't keep the
notification in history unless the "desktop-entry" hint is provided and
declares a valid desktop file. Unfortunately blobdrop does not ship a
desktop file (as it is not really a GUI app) and hardcoding another
foreign desktop file is bullcrap too, so for now we don't support
keeping the notification in the history, which is fine for now.