Allow to print paths as OSC8 hyperlinks
This prints the path directly to stdout via OSC8 just like "ls
--hyperlink=auto".
Terminal emulators may implement drag and drop functionality universally
for OSC8, so this can act as a third frontend.
Right now I don't know of any terminal emulator that actually does this,
but I might implement it for wezterm in the future.