Use string parameters for auto-quit option
Accepting {0,1,2} as parameters is as non-descriptive as it gets.
Improve on this usability problem by renaming the parameters to actual
strings that one is able to work with and that can be understood without
having to read the man page:
0 -> never
1 -> first
2 -> all

For now we accept both the old format and the new format, but this patch
already deprecates the old format and eventually we will want to only
accept the new string parameter format. But that will likely require a
major version bump.