Add immediate drag frontent
This is a very neat feature in X11 that allows you to automatically drag
and drop - the user only has to click on the target window.
I think it also works on MacOS, but I haven't tested it.

No more battling with the touchpad with this. :)

In the future we should probably unite all the different frontend
options into one single frontend option that can toggle between all
available frontends.