Add a demo for the immediate drop frontend
The idea was first mentioned in an dragon drag and drop issue [0].
And it was first implemented in clidrag [1], so we should give credits
to them.

Turns out, dragging and dropping is way more convenient when you don't
have to manually drag from a window, but it just starts automatically
and you don't even need to hold a mouse button.

[0] https://github.com/mwh/dragon/issues/57
[1] https://github.com/rkevin-arch/CLIdrag