Autohide the parent terminal if possible
"if possible" here means not using meme display servers, that don't
support this usecase for "security" reasons, unless you convince the
committee consisting of only the three largest desktop environments
(with at least one having very strong opinions) that this would be a
valid usecase to have and then proceed to write an RFC for a workaround
API (probably using at least three different workaround IPCs) in the
most cursed xml syntax, that you then are allowed to bikeshed over for
the coming years, until someone agrees that this trivial usecase that
has been working fine on the supposed deprecated display protocol, is a
valid usecase to have.
Yeah, I can't wait for the year of the Arcan desktop.

In all seriousness, xcb might not be the best documented API (and it
took quite some tinkering to figure this out), but still it's a pretty
reasonable design.

The feature that is implemented here, is to automatically keep the
parent terminal below other windows, because we probably don't want to
drag into the terminal window, that we just started this command from
and instead want to drop into another window, that might be currently
visually blocked by the terminal (which is pretty likely given that we
just launched blobdrop from a terminal).
After the drag we need to unhide the terminal window again.

My first thought when implementing this feature was trying to figure out
the connected pseudo-terminal of blobdrop and then figure out which
terminal is providing that pseudo-terminal.

Turns out this task is pretty much impossible, e.g. if we start tmux,
this information is completely lost.
For example running "pstree -slpa $$" in my current shell returns:

systemd,1
  └─systemd,664 --user
      └─tmux: server,5271
          └─zsh,87288
              └─pstree,87379 -slpa 87288

The information that wezterm is the underlying GUI terminal emulator
(and not tmux) is completely lost.

Well so that attempt is not working, let's find another way then:
As mentioned already, it is pretty likely that the terminal is the
currently focused window. Good, now we know what window to hide (and
that part is probably even implementable on Wayland already). The next
part is a bit more involved though:
We want to auto-hide the terminal window by setting the "Keep below
other windows hint". We can do this by adding the "_NET_WM_STATE_BELOW"
atom with xcb.
This would be equivalent to the following xdotool command:

xdotool windowstate --add BELOW "$(xdotool getactivewindow)"

For removing the "NET_WM_STATE_BELOW" flag afterwards, we would run:
xdotool windowstate --remove BELOW "$(xdotool getactivewindow)"

Any normal person would just start that command in a process and be done
with implementing that feature. I like to step into rabbit holes though.
Also xdotool uses xlib under the hood, implementing this with xcb feels
a lot cleaner.

According to the spec [0] we can't just use `xcb_change_property()`,
because that is only honored when the window is not already mapped.
Instead we have to send a message, e.g. a `xcb_client_message_event_t`
with `xcb_send_event()` [1].
It's much easier than that though, as xcb_ewmh maps this usecase in a
convenience function in util-wm [2] with
`xcb_ewmh_request_change_wm_state()`.
It is then important that we commit that state change by getting the
reply back with `xcb_get_window_attributes_reply()` and then flush our
messages with `xcb_flush()`.

As a comparison here is how Qt sets those flags [3], but it can
obviously use `xcb_change_property()` because it still has to map the
window (which is not possible in our case).

For establishing the xcb connection we try to reuse the Qt connection,
otherwise we establish a new one.

Something worth improving is if the user sends ^C, the program
terminates without removing the hint again, so the window stays below
other windows indefinitely. Maybe we could catch the signal to improve
this.

We should probably refactor the duplicate connection code into an extra
class in the future too, but for now I don't really care.

Right now this autohide logic only takes effect with the immediate
frontend, it might make sense to enable this for more backends in the
future.
We should also reconsider linking against libxcb directly, this will
annoy Wayland illusionists for sure. We can solve this by resolving the
symbols at runtime.

[0] https://specifications.freedesktop.org/wm-spec/1.3/ar01s05.html
[1] https://stackoverflow.com/a/68038842
[2] https://tudorr.ro/res/docs/xcb/util-wm/group__xcb____ewmh__t.html#ga706e1eaa15fd71e6c2da24a1559a194c
[3] https://codebrowser.dev/qt6/qtbase/src/plugins/platforms/xcb/qxcbwindow.cpp.html#1080