Use new std::chrono overload for timers
This requires Qt 6.6

Qt 6.6 also requires a nice change that allows the source files for QML
files to be stripped entirely from the binary [0] [1].

To use that, we could mark the qt_add_qml_module() call with
QT_DISCARD_FILE_CONTENTS, but that would completely remove ABI
compatibility, as the QMLEngine can only load QML binaries, that were
compiled with the same Qt version that the user is running. That would
be kinda broken, so we do not enable this feature for now.

[0] https://bugreports.qt.io/browse/QTBUG-103481
[1] https://codereview.qt-project.org/c/qt/qtbase/+/446176