Handle Unix signals
This fixes a usability problem, when blobdrop sets the keep-below hint,
but then is forcefully terminated by the user pressing ^C, which means
that the keep-below hint is never lifted.

Fix this by catching terminating Unix signals and then performing a
last-second cleanup before actually quitting.

Catching Unix signals is trivial [0] if we use only async-signal-safe
methods [1]. Howver we acually need to do some Qt function calls, that
would definitely violate these restrictions, so instead we use a more
elaborate setup [2], where we perform some safe writes via Unix sockets
during the signal interception, which will later trigger a Qt signal on
a listening QSocketNotifier when we are out of the critical zone again.
When the QSocketNotifier::activated() triggers, we are then safe to do
any operations we like.

[0] https://gist.github.com/azadkuh/a2ac6869661ebd3f8588
[1] https://pubs.opengroup.org/onlinepubs/000095399/functions/xsh_chap02_04.html#tag_02_04_03
[2] https://doc.qt.io/qt-6/unix-signals.html