Quit when pressing ESC in terminal
Holy cow, why are POSIX terminals such a shitshow???

First of all, changing the terminal mode with tcsetattr() is not reset
upon termination [0], so we have to manually make sure that we reset it
back to what it was beforehand by storing the original state with
tcgetattr().

Of course if our program is killed, there is nothing we can do to
restore it.

Then what the hell is wrong with canonical vs non-canonical mode?
We can't use canonical mode, because parsing line-by-line we could not
react to ESC right away.
So we are forced to use non-canonical mode, but then the fun really
begins...

If we set VMIN to 0, then std::cin.get() closes the whole stream as soon
as no input is not available right away, even though the user might type
something in the future, so that is completely broken.

If we set VMIN to 1 (which we are doing now), then std::cin behaves
correctly again, but the get() blocks now (expected), so that means we
resort back to getting one interrupt for every single letter and we are
no longer able to parse a line in one go (or in a for loop), but have to
get a socket "activated" event for ever char.

The termios(3) man page is very helpful to read for this too.

[0] https://stackoverflow.com/questions/47850223/should-i-reset-termios-settings-on-sigint-sigterm