(WIP) Remove xcb hard-dependency
I am not sure if this is the right thing to do, most people have xcb
still installed, even on Wayland, so this doesn't really solve a problem
at the moment.

Maybe something to look into in the future.

Also I don't really care that much about Wayland, so not sure why I am
even doing this...
The xcb API usage in this program actually allows for pretty neat UX so
if anything we should force users to have xcb.
Let's scratch this WIP and stash it in another branch, if we ever
redecide in the future...