Disable immediate frontend on Wayland
This changes the default frontend on Wayland to always be the gui
frontend.

As usual, Wayland is in the way of good usability and does not have
support for immediate drag operations (and probably never will), as the
spec mandates that `wl_data_device::start_drag()` is passed the serial
number of an implicit grab [0], and implementations adhering to the spec
explicitly check for such an active implicit grab [1].

This is kinda broken and does not allow us to start the drag right away,
meaning on Wayland we have to live with this inferior usability.
The same problem was also discovered in clidrag [2].

Funnily enough, it works if you use XWayland (although you can then only
drop into other XWayland applications).

You know your display protocol is utterly broken, when the compatibility
layer implementation for the display server you are supposedly replacing
has more features and usability than the native implementation.
But I am sure, the Wayland folks will come up with a replacement
workaround API soon, that they can bikeshed over for years to come...

[0] https://wayland.freedesktop.org/docs/html/apa.html#protocol-spec-wl_data_device
[1] https://invent.kde.org/plasma/kwin/-/blob/754b549f01b1adee727aa8e155e08cda8af91f44/src/wayland/datadevice.cpp#L99
[2] https://github.com/rkevin-arch/CLIdrag/issues/1