Add unit test for path URIs
This is good to ensure that we don't regress this in the future.

Right now the building of the unit tests is not optimal, as we rebuild
all source files for the tests again. So if we build with
-DBUILD_TESTING=ON, then we build each source file twice.

We should solve this in the future, by splitting them off into a shared
library.

Something cursed is, that the enable_testing() command **has** to be in
the top-level CMakeLists [0], putting it into the tests subdirectory
will not work (aka the tests will not be visible to ctest).

The unit tests are also run on the CI.

[0] https://stackoverflow.com/a/30264765