Disable QML compilation
Unfortunately Qt has no ABI-stability regarding compiled QML code.
This means that QML code compiled with one version of Qt will not run
anymore when Qt is upgraded.

I consider this totally broken behaviour and thus disable the
compilation for now. Maybe the ABI stability will be fixed soon.

The QML-compiler is split in three parts:
- qmltc
- qmlcachegen
- qmlsc

The latter one is only available for commercial licenses [0], so we
don't have to bother deactivating it.
qmltc is not enabled by default [1], we would need to explicitly set the
ENABLE_TYPE_COMPILER flag [2], which we don't do anyway.

However, qmlcachegen is enabled by default so we have to disable it with
"--only-bytecode" [3].

[0] https://doc.qt.io/qt-6/qtqml-qtquick-compiler-tech.html
[1] https://doc.qt.io/qt-6/qt-add-qml-module.html#compiling-qml-to-c-with-qml-type-compiler
[2] https://doc.qt.io/qt-6/qt-add-qml-module.html#enable-type-compiler
[3] https://doc.qt.io/qt-6/qt-add-qml-module.html#caching-compiled-qml-sources