Revert "gui: Use thumbnail as drag target"
This reverts commit 8f3e6290c32ac3bdefd7ddf867e6aca874749ac1.

Unfortunately it seems that there is a Qt bug with HighDPI that causes
the drag pixmap to be blurry, if the scaling factor is something else
than 1.
We can reapply the patch once this issue is fixed upstream.

Until then we continue using the icon as a thumbnail instead of the
image directly.