Add an intercept option
This option allows to intercept external drag and drop events and
converts them into drag and drop operations using any of the available
blobdrop frontends.

This can be useful for touchpad users to allow converting a drag and
drop event that requires holding the mouse button into one that does
not.
Binding this to a shortcut can improve UX for touchpad users
substantially.