Allow to shorten frontend to any prefix
This means that users will now be able to type e.g. "blobdrop -f clip"
instead of the full "blobdrop -f clipboard".

Right now the implementation has to be cut short a bit because of
missing std::ranges::to (will be available in gcc 14 and clang 17),
which means that we are not able to simply convert to an std::vector so
that we could check that the size is equal to 1. So for now, we only
check if it is not empty, which means it might not be unique.

We should fix this once gcc 14 is out.