Implement support for ssh remote transparency
This idea was originally meant only as a shitpost [0], but it can be
quite a nice feature in some niche usecases.

Implementation-wise the only finicky thing is that it is hard to get
reliable heuristics to determine the remote URI.
Sometimes it is even impossible to do so, e.g. when forwarding over
multiple hops. Maybe we should consider adding a "-P/--prefix" option in
the future to allow overwriting the prefix of the URI, in case the
heuristic fails.

For now this should be working good enough for most setups though.

The get_local_domain implementation is a bit hacky at the moment, e.g.
it does not cache if e.g. getaddrinfo() fails, only if it succeeds.
So if it's failing reliably, then it will call it for every single path
again.
But we have to leave some good stuff to be implemented for later. :)

Also it is unclear if sometimes the hostname should be removed from the
FQDN, so that we always get a proper domain name only.

Note that this feature requires using trusted X11 forwarding, i.e. the
use of "ssh -Y" instead of "ssh -X".

Right now it also does not seem to work with the gui frontend, but the
immediate frontend is fine.

[0] https://www.reddit.com/r/linux/comments/17to5xc/comment/k8ztcxp/