Disable DropArea for our own drags
It doesn't make sense to drag our own items back into the app itself in
sink mode and it only creates a confusing animation, when starting to
drag items.

The animation is still enabled when dropping.

It is a shame that direct QML bindings do not work inside into another
ListView.

An example of the confusing animation can be seen as an unrelated
sideeffect in the demo of #5.