Do not use immediate mode by default on MacOS
According to #7 the immediate frontend does not work on MacOS, so we
should probably not use it by default.

It can still be forced with "-f immediate", because I am not 100% sure
yet, if maybe someone can get it to work on MacOS. However, if we get
further reports that this is impossible to implement on MacOS (like on
Wayland), then we should also disable the forcing flag (like we already
do for Wayland).

Note that we do not change the ifdef to Q_OS_LINUX, because we do not
want to exclude the BSDs.