nix: Test both gcc and clang
Unfortunately it is impossible to run both check steps individually [0],
which would have allowed us to map the concurrent builds to the CI job
level.
Now this way it will be a bit confusing as both jobs will run
concurrently (but in the same CI job) and their stdout will race against
each other for maximum confusion, so the following output is possible:

```
The C compiler identification is GNU 14.2.0
The C compiler identification is Clang 18.1.8
The CXX compiler identification is GNU 14.2.0
The CXX compiler identification is Clang 18.1.8
```

And thus it will also not be possible to directly see which of the two
caused a compiler error.
Fortunately I do not care too much as it is easy for me to run the build
locally for just one of both compilers.

[0] https://github.com/NixOS/nix/issues/8881