Allow to show only icons in the gui window
This brings us more close to dragon compatibility. Note that the command
line argument is not completely compatible, the "-i" flag is already
taken, so we have to opt for "-I" instead. The long-form argument
"--icon-only" is identical at least.

The implementation is a bit hacky, since we manually adapt the size of
the window for now if the flag is present. Ideally the layout would
automatically dictate the size in a bottom-up way.
In the future we can also add an option to use the original image size
as dimensions for the thumbnail. Perhaps "--thumb-size 0" could be used
for this purpose. Right now depending on the image sizes, the default
thumbnail size of 64 can be a whacky in combination with this flag.

Closes #10