Adopt recommended QML module file layout
This implements the recommended file layout similar to commit
e0633ff8cece18852e5d5ce9eb6f597bad589e60 in quartz, except this time we
try it a bit different to the recommended file structure [0].

We still keep the file structure following the URI of the QML module,
but instead of creating a separate CMakeLists.txt inside the inner most
directory, we create everything from the root CMakeLists.txt.

This still seems to work, but a quick look at the resources with
Gammaray suggests that the source QML path structure is now redundantly
duplicated for all QML resources. This can possibly be fixed by setting
the resource output path for the QML module, but it is not really
important for QML files that only we consume.

Otherwise if we would use a second CMakeLists.txt inside the QML module
path, we would also need to move the entire qt_add_executable() call to
that CMakeLists.txt file as well to avoid weird errors with
"debug_metatypes.json:: Failed to parse JSON: 5 illegal value" [1] [2].

[0] https://doc.qt.io/qt-6/qt-add-qml-module.html#separate-backing-and-plugin-targets
[1] https://bugreports.qt.io/browse/QTBUG-122702
[2] https://bugreports.qt.io/browse/QTBUG-123574